requirements.txt
pymongo
dnspython3
requests